# docs/stream-tide.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "stream-tide",
    "description": "",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/v1/videos/upload": {
      "post": {
        "summary": "Get videos",
        "x-apidog-folder": "Video",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Video"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "test title"
                  },
                  "description": {
                    "type": "string",
                    "example": "abc def efg"
                  },
                  "file": {
                    "type": "string",
                    "example": "file:///home/shohanur/Personal-Projects/stream-tide/testing_videos/2.webm",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9529451-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/videos/published": {
      "get": {
        "summary": "Get published videos",
        "x-apidog-folder": "Video",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Video"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": false,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "title": {
                            "type": "string"
                          },
                          "viewCount": {
                            "type": "integer"
                          },
                          "thumbnailUrl": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "visibility": {
                            "type": "string"
                          },
                          "owner": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "firstName": {
                                "type": "string"
                              },
                              "lastName": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "firstName",
                              "lastName"
                            ],
                            "x-apidog-ignore-properties": [],
                            "x-apidog-orders": [
                              "id",
                              "firstName",
                              "lastName"
                            ]
                          }
                        },
                        "x-apidog-ignore-properties": [],
                        "x-apidog-orders": [
                          "id",
                          "title",
                          "viewCount",
                          "thumbnailUrl",
                          "createdAt",
                          "visibility",
                          "owner"
                        ]
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "currentPage": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "totalItems": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "currentPage",
                        "pageSize",
                        "totalItems",
                        "totalPages"
                      ],
                      "x-apidog-ignore-properties": [],
                      "x-apidog-orders": [
                        "currentPage",
                        "pageSize",
                        "totalItems",
                        "totalPages"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data",
                    "pagination"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data",
                    "pagination"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9542164-run",
        "security": []
      }
    },
    "/api/v1/videos/{videId}": {
      "get": {
        "summary": "Get video details",
        "x-apidog-folder": "Video",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Video"
        ],
        "parameters": [
          {
            "name": "videId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "80",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "viewCount": {
                      "type": "integer"
                    },
                    "hlsVideoUrl": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "visibility": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "ownerId": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "description",
                    "viewCount",
                    "hlsVideoUrl",
                    "createdAt",
                    "visibility",
                    "status",
                    "ownerId"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "id",
                    "title",
                    "description",
                    "viewCount",
                    "hlsVideoUrl",
                    "createdAt",
                    "visibility",
                    "status",
                    "ownerId"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9542175-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/videos/published/{videId}": {
      "get": {
        "summary": "Get public video details",
        "x-apidog-folder": "Video",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Video"
        ],
        "parameters": [
          {
            "name": "videId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "80",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "title": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "viewCount": {
                          "type": "integer"
                        },
                        "hlsVideoUrl": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "visibility": {
                          "type": "string"
                        },
                        "owner": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer"
                            },
                            "firstName": {
                              "type": "string"
                            },
                            "lastName": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "firstName",
                            "lastName"
                          ],
                          "x-apidog-ignore-properties": [],
                          "x-apidog-orders": [
                            "id",
                            "firstName",
                            "lastName"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "title",
                        "description",
                        "viewCount",
                        "hlsVideoUrl",
                        "createdAt",
                        "visibility",
                        "owner"
                      ],
                      "x-apidog-ignore-properties": [],
                      "x-apidog-orders": [
                        "id",
                        "title",
                        "description",
                        "viewCount",
                        "hlsVideoUrl",
                        "createdAt",
                        "visibility",
                        "owner"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9542183-run",
        "security": []
      }
    },
    "/api/v1/videos/{videoId}/make-private": {
      "put": {
        "summary": "Make video private",
        "x-apidog-folder": "Video",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Video"
        ],
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9542184-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/videos/published/{videoId}": {
      "put": {
        "summary": "Update video metadata",
        "x-apidog-folder": "Video",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Video"
        ],
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apidog-ignore-properties": [],
                "x-apidog-orders": []
              },
              "example": {
                "title": "abc 1234",
                "description": "Abcd efg"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "null"
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9542186-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/comments/videos/{videoId}": {
      "get": {
        "summary": "Get video comments",
        "x-apidog-folder": "Comment",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "name": "videoId",
            "in": "path",
            "description": "",
            "required": true,
            "example": "123",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "example": "1",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": false,
            "example": "10",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "content": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string"
                              },
                              "updatedAt": {
                                "type": "string"
                              },
                              "user": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "integer"
                                  },
                                  "firstName": {
                                    "type": "string"
                                  },
                                  "lastName": {
                                    "type": "string"
                                  }
                                },
                                "required": [
                                  "id",
                                  "firstName",
                                  "lastName"
                                ],
                                "x-apidog-ignore-properties": [],
                                "x-apidog-orders": [
                                  "id",
                                  "firstName",
                                  "lastName"
                                ]
                              }
                            },
                            "x-apidog-ignore-properties": [],
                            "x-apidog-orders": [
                              "id",
                              "content",
                              "createdAt",
                              "updatedAt",
                              "user"
                            ]
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "currentPage": {
                              "type": "integer"
                            },
                            "pageSize": {
                              "type": "integer"
                            },
                            "totalItems": {
                              "type": "integer"
                            },
                            "totalPages": {
                              "type": "integer"
                            }
                          },
                          "required": [
                            "currentPage",
                            "pageSize",
                            "totalItems",
                            "totalPages"
                          ],
                          "x-apidog-ignore-properties": [],
                          "x-apidog-orders": [
                            "currentPage",
                            "pageSize",
                            "totalItems",
                            "totalPages"
                          ]
                        }
                      },
                      "required": [
                        "data",
                        "pagination"
                      ],
                      "x-apidog-ignore-properties": [],
                      "x-apidog-orders": [
                        "data",
                        "pagination"
                      ]
                    }
                  },
                  "required": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ],
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": [
                    "statusCode",
                    "message",
                    "success",
                    "data"
                  ]
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/642230/apis/api-9542232-run",
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Develop Env"
    }
  ]
}